bundle nodes are now treated as floats

+++ fileio.py
@@ -36,7 +36,8 @@ class InpFileParser(object):
         nodes = re.split("\s+|,\s*", config.get("Bundle", "nodes"))
         nodes = filter(None, nodes)
         #nodes.reverse()
-        self.parent.nodes = map(int, nodes)
+        #self.parent.nodes = map(int, nodes)
+        self.parent.nodes = map(float, nodes)
         if len(self.parent.nodes) != len(self.parent.caxfiles):
             print "Error: Invalid node list."
             return False
@@ -54,7 +55,8 @@ class InpFileParser(object):
             btf_nodes = re.split("\s+|,\s*", config.get("BTF", "nodes"))
             btf_nodes = filter(None, btf_nodes)
             #btf_nodes.reverse()
-            self.parent.btf_nodes = map(int, btf_nodes)
+            #self.parent.btf_nodes = map(int, btf_nodes)
+            self.parent.btf_nodes = map(float, btf_nodes)
         else:
             self.parent.btf_nodes = self.data.nodes
